u64 to u32

+++ bee-ledger/src/types/snapshot/milestone_diff.rs
@@ -69,14 +69,14 @@ impl Packable for MilestoneDiff {
             }
         }
 
-        (self.created_outputs.len() as u64).pack(packer)?;
+        (self.created_outputs.len() as u32).pack(packer)?;
 
         for (output_id, created) in self.created_outputs.iter() {
             output_id.pack(packer)?;
             created.pack(packer)?;
         }
 
-        (self.consumed_outputs.len() as u64).pack(packer)?;
+        (self.consumed_outputs.len() as u32).pack(packer)?;
 
         for (output_id, (created, consumed)) in self.consumed_outputs.iter() {
             output_id.pack(packer)?;
@@ -118,7 +118,7 @@ impl Packable for MilestoneDiff {
             None
         };
 
-        let created_count = u64::unpack::<_, VERIFY>(unpacker).coerce()?;
+        let created_count = u32::unpack::<_, VERIFY>(unpacker).coerce()?;
         let mut created_outputs = HashMap::with_capacity(created_count as usize);
 
         for _ in 0..created_count {
@@ -128,7 +128,7 @@ impl Packable for MilestoneDiff {
             created_outputs.insert(output_id, created_output);
         }
 
-        let consumed_count = u64::unpack::<_, VERIFY>(unpacker).coerce()?;
+        let consumed_count = u32::unpack::<_, VERIFY>(unpacker).coerce()?;
         let mut consumed_outputs = HashMap::with_capacity(consumed_count as usize);
 
         for _ in 0..consumed_count {
